[base] Split string on space and comma instead of using lodash.words for tokenization (#1222)
In #1168 we used lodash.words for tokenization of the search string. Can't really remember the rationale behind it (think it had something to do with unicode, but in retrospect it feels misguided). Using lodash.words had some unintended side effects, like splitting words like 2pac into ['2', 'pac'].

This PR replaces lodash.words with a simple regex that splits the search string on space, comma and dash.