[validation] Fix a performance issue when validating blocks
This changes the validation of a block to use it's own validation function. Earlier it used the 'custom' function which spins up one or two setTimeouts for each block. When working on huge documents that was too much for the browser (it seems, in Node it runs pretty good). This commit changes it to not use setTimeouts at all (if not the validation is overridden with an async validation function).