[base] Switch to a warning instead of hard failure on missing tracker context (#2059)
When using the `useReporter()` or `useReportedValues()` hook outside of a Tracker element context we would previously fail hard. Since this is currently used for non-critical features like the change connectors and presence we should hard fail when it happens. This PR issues a console warning instead of crashing the studio. A future improvement would be to make the failure mode configurable, since there might be situations where it's used for more critical features.