fix(form-builder): temporarily re-export moved patches modules
For some reason, after upgrading from 2.x to 2.6.0, there is an error where
`lib/sanity/uploads/uploaders.js` is referencing `../../utils/patches` instead
of the expected `../../patch/patches`. This isn't actually the contents in the
tarball, so it seems to somehow be an issue with yarn? To (try) to work around
this, we're re-exporting the patches module in the old location for now.