fix: preserve focus on calendar buttons between renders
this fixes an issue with the date picker that would sometimes make it lose focus when navigating from the last day of a month to the first day of the next. This would only happen when moving from the last day of a month where the calendar grid displayed 6 weeks to the first day of a month that would display as 5 weeks (e.g. may 2021 to june 2021).