chore: remove strict folder overrides (for now)
Using separate tsconfig files on subdirectories didn't work all that well with composite/project references as it was currently set up. It messed with mapping from dist imports to source files / d.ts, which required either a restart of the ts language server or a recompile for changes to be reflected.