chore: add verify monorepo dependency script + run on postinstall
This adds a script that verifies that monorepo packages depends on other monorepo packages at *their current versions*. E.g. if `@sanity/desk-tool`'s package.json depends on a version of `@sanity/base` which is not the version `@sanity/base` is currently on, you will now get an error on `npm install`, telling you to fix the version discrepancy.

This helps in recovering from a problem that happens quite frequent when working in a feature branch where you add a monorepo package as a dependency to another monorepo package. If a new version of the newly added dependency has been released and you're doing a rebase against main, the version of the added dependency falls behind, usually undetected. This again is causing strange and disruptive behavior like whole packages getting deleted from disk, etc.

A nice future improvement would be to offer a way to fix this issue automatically (e.g. by adding --fix)