fix(cli): clear auth token on 401 on logout
When attempting to log out of the CLI, one might encounter situations where the
session is either timed out, has been cleared or does not exist. In these cases
it is perfectly acceptable to just clear the token from the CLIs point of view,
instead of throwing an error which is not actionable.