fix(datastores): merge the autocommit stream with the operation results instead of side-effectsy subscribe
We currently see an issue with multiple identical mutations being submitted when changing a document after hot reload. This happens because whenever the studio wants to submit a patch through the buffered document/mutator it emits the patch to a global shared stream (subject) which is handled downstream , and whenever we want to subscribe to operation results we cache this on the client instance and start subscribing in order to support autocommit. The fix is to merge the autocommit$ stream with the operation results stream so that it is only subscribed for as long as the operation results stream is subscribed for.

A potential issue with this is that changes to the document  will not be committed unless something in the runtime is also still subscribing to the operation results stream (up until now it was enough that it subscribed at some point in time because the subscription were never unsubscribed once set up. In practice - at least  in today’s studio - this should probably not cause any big issues as there will normally be something in the view layer having a subscription to the operation results for any edited document (e.g. to show notifications when the document get’s published, etc), but this definitely something to be aware of.