feat(form): introduce spreadable "elementProps" to reduce custom input boilerplate
This passes a spreadable `elementProps` property to all input components. This currently includes things like `onFocus`/`onBlur` event handlers, `ref`, `value` (for primitive inputs) and an `onChange` that handles "native" events and emits the correct patches, and the idea is that you can safely pass it on to any "native" input element. This reduces the amount of boilerplate needed for custom inputs (especially those dealing with primitive values). For inputs dealing with complex values (objects and arrays), `value` is not included in `elementProps`, since it doesn't make sense to pass it to any native input.