fix(form): respect enableHiddenCheck condition in formState
Return null here only when enableHiddenCheck, or we end up with array members that have 'item: null' when they
really should not be. One example is when a block object inside the PT-input have a type with one single hidden field.
Then it should still be possible to see the array member item, even though all of it's fields are null.