(form/inputs): PT: fix various path and selection issues (#4310)
* fix(portable-text-editor): use editor relativePath for PTE API methods

The PTE would not know about the full studio paths, only relative to the editor itself. Use these when calling .delete

* fix(form/inputs): fix scroll to focused member issues

Make sure that we make a selection in useScrollToOpenendMember only
when the actual focus is directly on the block itself.

Whenever there was focus on a inlineblock or annotation inside that block this
selection was incorrectly made too.

Also PortableTextInput was incorrectly setting the wrong focus path (relative instead of full).
This is now fixed.

* refactor(form/inputs): make inlineObject render without member

To get immediate response when copying or moving inlineObjects, render the object even though the form member isn't created yet.

* fix(form/inputs): prevent toolbar background taking focus from the editor

This still needs to be here or clicking somewhere in the toolbar will remove the focus from the editor when it shouldn't

* refactor(form/inputs): set focus again, when the editor is re-rendered

If the PT-input hasFocus, and the Editor comp. is re-rendered, re-apply focus to the editor.
This now works generally, so remove spesific full-screen-toggle focusing code.

* fix(portable-text-editor): remove unnecessary selection change

This will not be taken care of by the selection plugin. Remove so we don't emit double.

* fix(form/inputs): fix issues with block style application and PT-input focus

This will fix and issue where focus was not set back to the editor after
changing a block's text style in the PT-input.

Remove complex logic that now can be simplified. We no longer need to track change here.

* refactor(form/inputs): remove unnecessary delete handler

This can now be called directly as there is not need to prevent the default.

* refactor(form/inputs): remove unnecessary tab handling

This tab handling no longer seems needed as the popover button will now take focus properly when the object is focused in the editor.