fix(form/inputs): fix Safari issue with annotation edit popovers being hidden
There is a regression for Safari that hides the annotation edit popover when clicking to edit it. This is because Safari creates a new text range that
isn't created by other browsers when clicking a button. This range is without focusNode and anchorNode and the popover is closed.

Fix this by requiring a focusNode to be set in the range before hiding the popover.