fix(form/inputs): improve tracking of formState focusPath for PT-input (#4466)
* fix(desk): DocumentPane must be keyed on id

The document id must be part of the key here, or focusPath isn't propery reset when changing documents.

* fix(form/inputs): improve tracking of formState focusPath for PT-input

This will improve tracking of the formState focus in the PT-input. It can now restore focus on single spans of text.
Also improved how we restore the editor selection when toggling fullscreen.

Renamed hook from useScrollToOpenendMember to useTrackFocusPath as it covers more what it does.

Made some small adjustments so that we scroll to both object blocks and text spans.

* fix(portable-text-editor): emit selection after focused and unchanged

If the selection is the same as when focus was lost, the editor will not emit a new one
because there is no onChange done to the editor instance.

Explicitly emit the selection here in that case.