Add boundary provider for nested content inside PT-input dialogs (#4486)
* chore(test-studio): add an author references block to simpleBlock

* fix(form/inputs): add boundary element provider for nested dialogs inside the PT-input

Wrap PT-input dialog content in a own boundary element provider, so that content inside can
use the actual content as the boundary element. This is needed to correctly calculate
the height of it for virtual array lists which will use the closest boundary for this.

Using state here instead of refs based on team experience with dialogs for the arrays.

---------

Co-authored-by: Robin Pyon <robin.pyon@gmail.com>