refactor(core/form/inputs): refactor focus handling in the PT-input
There were a combination of issues with how the focusPath is tracked
inside the PT-input and applied from and to the editor.

This will improve how the PT-input responds to outside changes
in focusPath and how it handles focus when deleting nodes etc.

There is also a fix for autoFocusing dialogs and popovers
for embedded PT-content.

fix(core/form/inputs): fix reference elm issue with PT-input

When creating new annotations the memberItem will not we available on the first render.
Use a local reference element through state instead to ensure it's always a reference elm for the popover.

fix(form/inputs): only scroll and select focusPath when needed

This will make sure we only scroll and set a selection to track the focusPath if the
selected block is different than what currently selected in the editor.

Also we must perform two scrollIntoViews, both for the boundaryElement
but also for the element inside the boundary element.

refactor(form/inputs): call onPathFocus through transition

There are performance issues with calling onPathFocus on the editor selection directly.
Previously this was mitigated with using a debounce on it. This does however have some
implications for the functions that is also tracking the focusPath. Where the
useTrackFocusPath hook sometimes catches up on a debouced onPathFocus call.

By instead calling onPathFocus in a React transition, we do
get the satisfactory performance optimization, but also the correct synchronization
of the focusPath state.

fix(form): set focus according to focus path on inputs inside block objects

fix(form/inputs): only focus the editor when absolutely needed

* Should be true only when either a single block is selected, or one of it's children are.
* When toggling fullscreen, we don't need to re-set the selection. Only focus or blur.
* We don't need an initial selection to the editor. It will be properly set according to the focusPath through useTrackFocusPath hook.

fix(form/inputs): make stricter rules for when to make a editor selection

In the PT-input, make sure we don't make a editor selection when the focusPath is pointing to a markDef (annotation) or are editing inside inline objects.

refactor(form/inputs): don't restore to initial selection

In the PT-input we should now go to the parent object instead. This is now taken care of by the individual components for annotation, blockObject and inlineObject instead.

refactor(form/inputs): improve how PT-input handles deletion of nodes

If deleting inline objects or block objects with the mouse through the menu,
we must prevent the mouse events first or focus gets lost.

Also safeguard that delete can only be clicked once by disabling the button.

refactor(form): restore original blockProps

Co-authored-by: Bjørge Næss <bjoerge@gmail.com>