fix(core/form/inputs): fix rendering regression for PT-input (#4530)
There was a rendering regression in 75dfb9ca1bbb6bc21cad9c8139e4a9d87f9522ad where nested PT-inputs (PT in PT) would not
properly update their form inputs.

Revert the use of refs for the input in the PT node components, and let the PortableTextInput figure that out instead.

I have added a condition that will only update the input if the member is open or it's value has changed.