refactor(portable-text-editor): improve onPaste typings (#4544)
* fix(portable-text-editor): fix incorrect path called for onPaste

This is supposed to be a keyed PortableText path (and not a Slate path)

* refactor(portable-text-editor): improve typings for onPaste fn

Using PortableTextBlock here is not correct, it should require a TypedObject only
from the outside, as all the rest will be taken care of.

Also create an own interface for the paste data so it can be easier read from the outside.